telegram-bot/requirements.txt
python-telegram-bot==21.10
firebase-admin==6.6.0
httpx==0.28.1
python-dotenv==1.1.0
apscheduler==3.11.0
